move around the callers of check-student and load-data
